moved the entityreader test and fixed some javadoc
