+++ utils/importexport/pom.xml
@@ -1,29 +1,46 @@
 <?xml version="1.0"?>
 <project
-        xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-        xmlns="http://maven.apache.org/POM/4.0.0"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>io.apicurio</groupId>
-        <artifactId>apicurio-registry</artifactId>
-        <version>2.0.0-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
-    </parent>
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+	xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+	<modelVersion>4.0.0</modelVersion>
+	<parent>
+		<groupId>io.apicurio</groupId>
+		<artifactId>apicurio-registry</artifactId>
+		<version>2.0.0-SNAPSHOT</version>
+		<relativePath>../../pom.xml</relativePath>
+	</parent>
 
-    <artifactId>apicurio-registry-utils-import-export</artifactId>
-    <packaging>jar</packaging>
-    <name>apicurio-registry-utils-import-export</name>
+	<artifactId>apicurio-registry-utils-import-export</artifactId>
+	<packaging>jar</packaging>
+	<name>apicurio-registry-utils-import-export</name>
 
-    <dependencies>
-    
-        <dependency>
-            <groupId>io.apicurio</groupId>
-            <artifactId>apicurio-registry-common</artifactId>
-        </dependency>
-    
-    </dependencies>
-    
-    <build>
-    </build>
+	<dependencies>
+
+		<dependency>
+			<groupId>io.apicurio</groupId>
+			<artifactId>apicurio-registry-common</artifactId>
+		</dependency>
+
+		<!-- Test Dependencies -->
+		<dependency>
+			<groupId>org.junit.jupiter</groupId>
+			<artifactId>junit-jupiter</artifactId>
+			<scope>test</scope>
+		</dependency>
+
+	</dependencies>
+
+	<build>
+		<testResources>
+			<testResource>
+				<directory>src/test/resources</directory>
+				<filtering>true</filtering>
+			</testResource>
+			<testResource>
+				<directory>src/test/resources-unfiltered</directory>
+				<filtering>false</filtering>
+			</testResource>
+		</testResources>
+	</build>
 </project>
