+++ utils/importexport/src/main/java/io/apicurio/registry/utils/impexp/ArtifactRuleEntity.java
@@ -29,7 +29,7 @@ public class ArtifactRuleEntity extends Entity {
     public String configuration;
 
     /**
-     * @see io.apicurio.registry.utils.impexp.storage.impexp.Entity#getEntityType()
+     * @see io.apicurio.registry.utils.impexp.Entity#getEntityType()
      */
     @Override
     public EntityType getEntityType() {
