+++ utils/importexport/src/main/java/io/apicurio/registry/utils/impexp/ArtifactVersionEntity.java
@@ -44,7 +44,7 @@ public class ArtifactVersionEntity extends Entity {
     public long contentId;
 
     /**
-     * @see io.apicurio.registry.utils.impexp.storage.impexp.Entity#getEntityType()
+     * @see io.apicurio.registry.utils.impexp.Entity#getEntityType()
      */
     @Override
     public EntityType getEntityType() {
