+++ utils/importexport/src/main/java/io/apicurio/registry/utils/impexp/ContentEntity.java
@@ -30,7 +30,7 @@ public class ContentEntity extends Entity {
     public byte[] contentBytes;
 
     /**
-     * @see io.apicurio.registry.utils.impexp.storage.impexp.Entity#getEntityType()
+     * @see io.apicurio.registry.utils.impexp.Entity#getEntityType()
      */
     @Override
     public EntityType getEntityType() {
