+++ utils/importexport/src/main/java/io/apicurio/registry/utils/impexp/GlobalRuleEntity.java
@@ -27,7 +27,7 @@ public class GlobalRuleEntity extends Entity {
     public String configuration;
 
     /**
-     * @see io.apicurio.registry.utils.impexp.storage.impexp.Entity#getEntityType()
+     * @see io.apicurio.registry.utils.impexp.Entity#getEntityType()
      */
     @Override
     public EntityType getEntityType() {
