+++ utils/importexport/src/main/java/io/apicurio/registry/utils/impexp/ManifestEntity.java
@@ -31,7 +31,7 @@ public class ManifestEntity extends Entity {
     public String exportedBy;
 
     /**
-     * @see io.apicurio.registry.utils.impexp.storage.impexp.Entity#getEntityType()
+     * @see io.apicurio.registry.utils.impexp.Entity#getEntityType()
      */
     @Override
     public EntityType getEntityType() {
