+++ utils/importexport/src/test/java/io/apicurio/registry/utils/impexp/EntityReaderTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.apicurio.registry.rest.v2.impexp;
+package io.apicurio.registry.utils.impexp;
 
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
@@ -23,34 +23,26 @@ import java.util.zip.ZipInputStream;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import io.apicurio.registry.AbstractRegistryTestBase;
-import io.apicurio.registry.utils.impexp.ArtifactRuleEntity;
-import io.apicurio.registry.utils.impexp.ArtifactVersionEntity;
-import io.apicurio.registry.utils.impexp.ContentEntity;
-import io.apicurio.registry.utils.impexp.Entity;
-import io.apicurio.registry.utils.impexp.EntityReader;
-import io.apicurio.registry.utils.impexp.GlobalRuleEntity;
-
 /**
  * @author eric.wittmann@gmail.com
  */
-class EntityReaderTest extends AbstractRegistryTestBase {
+class EntityReaderTest {
 
     /**
      * Test method for {@link io.apicurio.registry.rest.v2.impexp.EntityReader#readEntity()}.
      */
     @Test
     void testReadEntity() throws Exception {
-        try (InputStream data = resourceToInputStream("../export.zip")) {
+        try (InputStream data = resourceToInputStream("export.zip")) {
             ZipInputStream zip = new ZipInputStream(data, StandardCharsets.UTF_8);
             EntityReader reader = new EntityReader(zip);
             Entity entity = null;
-            
+
             int contentCounter = 0;
             int versionCounter = 0;
             int globalRuleCounter = 0;
             int artyRuleCounter = 0;
-            
+
             while ( (entity = reader.readEntity()) != null ) {
                 if (entity instanceof ContentEntity) {
                     contentCounter++;
@@ -65,7 +57,7 @@ class EntityReaderTest extends AbstractRegistryTestBase {
                     globalRuleCounter++;
                 }
             }
-            
+
             Assertions.assertEquals(1003, contentCounter);
             Assertions.assertEquals(5, versionCounter);
             Assertions.assertEquals(1, artyRuleCounter);
@@ -73,4 +65,13 @@ class EntityReaderTest extends AbstractRegistryTestBase {
         }
     }
 
+    /**
+     * Loads a resource as an input stream.
+     * @param resourceName the resource name
+     */
+    protected final InputStream resourceToInputStream(String resourceName) {
+        InputStream stream = getClass().getResourceAsStream(resourceName);
+        Assertions.assertNotNull(stream, "Resource not found: " + resourceName);
+        return stream;
+    }
 }
